refactor(language): split test.rs into per-phase test modules
test.rs had grown to ~1568 lines covering the whole pipeline in one
file. Converted to a test/ module directory and split by pipeline
phase: lexer, diagnostics, parser, compiler, extended_grammar,
capability_binding, graph_materialisation, registry_binding,
resolver, and provenance_diff_testkit, with mod.rs keeping the shared
SUPPORT_AGENT fixture and module declarations. No behavior or
coverage change — same 76 tests, same assertions.
